Program "Inventory" from ATBSWP
Just added the code which I've written for the task mentioned in Automate the Boring Stuff with Python book by Al Sweigart. Love the book!
The task itself can be found in the book (I've actually purchased its printed version and proud of it!) and also on the official website (https://automatetheboringstuff.com/2e/chapter5/). I highly recommend both this book and its author, Al Sweigart.

NB: But I've made some changes: firstly I wanted to use the initial inventory dictionary, with the values already quoted. So, I've added the new list (loot) to the current inventory which was created. Secondly I've made this via PyCharm app and saved it as a .py file at first, and only then copied it to my Jupyter Notebook and saved it as .pynb file. Now it is here. 
(I've written two options of the code: with "try" and "except" methods - which is shorter and less time consuming, and without them, making a loop within a loop).
Hope you'll like the program and the code.